Various improvements and updates

+++ app/controllers/claims_controller.rb
@@ -10,23 +10,23 @@ class ClaimsController < ApplicationController
       @sort_msg=sort_bar("Claims","td")
     end
     if (params[:filter]=="r")
-      qry="claims.id in (SELECT claim_id FROM claim_reviews WHERE claim_reviews.review_sharing_mode=1 AND review_verdict!='')"
+      qry="claims.id in (SELECT claim_id FROM claim_reviews WHERE (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") AND claim_reviews.review_sharing_mode=1 AND review_verdict!='')"
       @filter_msg=filter_bar("Claims","r")
     elsif (params[:filter]=="u")
-      qry="claims.id in (SELECT claim_id FROM claim_reviews WHERE claim_reviews.user_id="+current_user.id.to_s+")"
+      qry="claims.id in (SELECT claim_id FROM claim_reviews WHERE (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") AND claim_reviews.user_id="+current_user.id.to_s+")"
       @filter_msg=filter_bar("Claims","u")
     elsif (params[:filter]=="n")
-      qry="NOT EXISTS (SELECT claim_id FROM claim_reviews WHERE claims.id=claim_reviews.claim_id AND ((claim_reviews.review_sharing_mode=1 AND claim_reviews.review_verdict!='') OR claim_reviews.user_id="+current_user.id.to_s+"))"
+      qry="(claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") AND NOT EXISTS (SELECT claim_id FROM claim_reviews WHERE claims.id=claim_reviews.claim_id AND ((claim_reviews.review_sharing_mode=1 AND claim_reviews.review_verdict!='') OR claim_reviews.user_id="+current_user.id.to_s+"))"
       @filter_msg=filter_bar("Claims","n")
     elsif (params[:q].present?)
       @filter_msg=filter_bar("Claims","a")
-      qry="lower(title) like lower('%"+params[:q]+"%') or lower(description) like lower('%"+params[:q]+"%') or lower(url_preview) like lower('%"+params[:q]+"%')"
+      qry=" (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and (lower(title) like lower('%"+params[:q]+"%') or lower(description) like lower('%"+params[:q]+"%') or lower(url_preview) like lower('%"+params[:q]+"%'))"
     elsif (params[:m].present?)
       @filter_msg=filter_bar("Claims","a")
-      qry="medium_id="+params[:m].to_s
+      qry=" (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and medium_id="+params[:m].to_s
     elsif (params[:s].present?)
       @filter_msg=filter_bar("Claims","a")
-      qry="src_id="+params[:s].to_s
+      qry=" (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and src_id="+params[:s].to_s
     elsif (params[:url].present?)
         @filter_msg=""
         output=""
@@ -49,11 +49,11 @@ class ClaimsController < ApplicationController
               imglist=imglist+"prev.onclick=function(){if(i==0){i="+i.to_s+";};document.getElementById('cimg').src=images[--i];}</script>"
               output=output+"\n"+imglist
               output=output+'<br><a href="#" id="prev">Change thumbnail</a><br><div id="final_url_preview" class="fragment">'
-              output=output+'<div style="text-align: left"><img src="'+preview.images.first.src.to_s+'" id="cimg" height=75 />'
+              output=output+'<div style="text-align: left"><img src="'+preview.images.first.src.to_s+'" id="cimg" style="max-width:128px;max-height:75px" />'
             elsif i==1
-              output=output+'<br><div id="final_url_preview" class="fragment"><div style="text-align: left"><img src="'+preview.images.first.src.to_s+'" id="cimg" height=75 /><br>'
+              output=output+'<br><div id="final_url_preview" class="fragment"><div style="text-align: left"><img src="'+preview.images.first.src.to_s+'" id="cimg" style="max-width:128px;max-height:75px" /><br>'
             elsif is_img(params[:url])
-              output=output+'<div id="final_url_preview" class="fragment"><div style="text-align: left"><img src="'+params[:url]+'" id="cimg" height=70 /><br>'
+              output=output+'<div id="final_url_preview" class="fragment"><div style="text-align: left"><img src="'+params[:url]+'" id="cimg" style="max-width:128px;max-height:75px" /><br>'
               titl=params[:url]
               desc="Image URL"
             else
@@ -76,12 +76,15 @@ class ClaimsController < ApplicationController
       else
         @filter_msg=filter_bar("Claims","a")
         if (params[:sort]=="r" or params[:sort]=="rp" or params[:sort]=="rn")
-          tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
+          tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
           @total_count=tmp.count.length
         elsif  (params[:sort]=="rt")
-          tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
+          tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
           @total_count=tmp.count.length
+        elsif !user_signed_in?
+            return
         else
+          if qry.nil? then qry="claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s; end
           tmp=Claim.where(qry).order("created_at DESC")
           @total_count=tmp.count
         end
@@ -89,12 +92,15 @@ class ClaimsController < ApplicationController
         return
       end
       if (params[:sort]=="r" or params[:sort]=="rp" or params[:sort]=="rn")
-        tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
+        tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
         @total_count=tmp.count.length
       elsif  (params[:sort]=="rt")
-        tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
+        tmp=Claim.joins(:claim_reviews).where("claims.id=claim_reviews.claim_id and (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+") and claim_reviews.review_sharing_mode=1 and claim_reviews.review_verdict!=''").group("claim_reviews.claim_id").order(sort_statement("claim",params[:sort]))
         @total_count=tmp.count.length
+      elsif !user_signed_in?
+          return
       else
+        if qry.nil? then qry="claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s; end
         tmp=Claim.where(qry).order("created_at DESC")
       end
       @total_count=tmp.count
@@ -105,7 +111,7 @@ class ClaimsController < ApplicationController
     @claims_msg=""
     @reviews_msg=""
     @warning_msg=""
-    dependent_reviews=ClaimReview.where("claim_id = ?",@claim.id).count("id")
+    dependent_reviews=ClaimReview.where("claim_id = ? and claim_id in (select id from claims where (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+")) ",@claim.id).count("id")
     if (dependent_reviews>0)
       @warning_msg="Deleting this record will also delete "
       @warning_msg=@warning_msg+" "+dependent_reviews.to_s+" dependent "+pl(dependent_reviews,"review")+".\n"
@@ -143,7 +149,7 @@ class ClaimsController < ApplicationController
   end
 
   def destroy
-    ClaimReview.where("claim_id = ?",@claim.id).destroy_all
+    ClaimReview.where("claim_id = ? and claim_id in (select claims.claim_id from claims where (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+"))",@claim.id).destroy_all
     @claim.destroy
     redirect_to root_path
   end
@@ -165,11 +171,11 @@ class ClaimsController < ApplicationController
     end
 
     def claim_params
-      params.require(:claim).permit(:id, :title, :medium_name, :src_name, :url, :description, :has_image, :has_video, :has_text, :url_preview)
+      params.require(:claim).permit(:id, :title, :medium_name, :src_name, :url, :description, :has_image, :has_video, :has_text, :sharing_mode, :url_preview)
     end
 
     def find_claim
-      @claim = Claim.find(params[:id])
+        @claim = Claim.where("id=? AND (claims.sharing_mode=1 OR claims.user_id="+current_user.id.to_s+")",params[:id]).first
     end
 
     def check_if_signed_in
